added file input and early revision of env save
System can now load a specified text file and will automatically attempt
parse and execute any text therein as Logo code. When exiting the
program using the 'exit' command or when using the new 'save' command,
any LNumbers and any procedures defined by a 'to... end' block will be
automatically saved to a text file called 'saveFile.txt,' and this file
is automatically loaded when the program starts, restoring the user's
previously defined words. The run tag arguments have been removed as
they will no longer be supported and are now largely obsolete: running
an REPL session and being able to load from file are no longer mutually
exclusive functions of the program, and the new interface allows for
more of this multiple functionality. Also added was a function whereby
the user can retrieve their last input by pressing the up arrow on the
keyboard, allowing quick re-entry or modification of the previous
command.