added text editor pane, large graphics display
Text editor pane now allows the user to write and save code to a
specified text file for future execution within the development
environment. A larger turtle graphics display pane is now available in
the graphics tab and switching to the graphics tab will hide the small
turtle graphics display.