text editor now supported by GUI tool bar
Improved text editor pane has been moved to a separate class,
TextEditor, and implemented within the existing user interface.
Additionally, the program now detects which directory the user is
running from and uses this as the base for the relative file paths for
URLs, rather than the hard-coded directory being used as a test bed for
the functionality of these elements. This allows the URL links for the
words printed with the 'help' command to function on any system. If
desired, the user can create an HTML document with the name [word].html
(replace '[word]' with the name of the word you would like to link to
this file) in the docs folder that will open when the user clicks that
word in the help list.

Known issues:
The window size resets to default when switching to or from the
'Graphics' tab.

Currently, switching to or from the 'Graphics' tab resets the graphics
pane, losing any
drawings and returning the drawing pen (the 'turtle') to the center of
the pane.