redesigned and improved graphics resize, doc gen
The large turtle graphics tab now displays a resized turtle graphics
pane that also saves the turtle's position, heading, color, and state
and preserves any lines drawn to the screen by the turtle. Lines that
are drawn off screen are currently being lost, so some drawings may be
cut off if they are too large to fit on the small turtle graphics pane
and the user switches to the small pane and back; a fix for this should
be coming shortly. Documentation is now automatically generated for all
system defined words via the 'genDocs', 'regenDocs', and 'help'
commands. Template documents are also generated for any user defined
procedures. Documentation is not yet finished for all words, but will be
coming in a future update.